+++ src/headless-darkmode.ts
@@ -1,4 +1,5 @@
-import { keycodeEquals } from "./utils";
+import { keycodeEquals } from "./utils"
+
 /**
  * Class for HeadlessDarkMode
  * @class
@@ -10,33 +11,40 @@ import { keycodeEquals } from "./utils";
  * @link toggleTheme
  * @link toggleThemeWithKeyboard
  * @link themeData
+ * @link observedAttributes
+ * @link attributeChangedCallback
+ * @link updateAllDateThemeAttributes
  * @since 0.3.1
  * */
 export class HeadlessDarkMode extends HTMLElement {
-  private button: HTMLButtonElement | HTMLElement | null;
+  private button: HTMLButtonElement | HTMLElement | null
+  private dataTheme: NodeListOf<HTMLElement>
+
   constructor() {
-    super();
-    this.button = this.querySelector('button')
-    this.dataset.theme = this.themeData
+    super()
+    this.button = this.querySelector("button")
+    if (!this.button.hasAttribute("aria-checked")) {
+      this.button.setAttribute("aria-checked", "false")
+    }
+    this.dataset.theme = this.themeData || "dark"
+    this.dataTheme = this.querySelectorAll("[data-theme]")
+    this.updateAllDateThemeAttributes(this.dataset.theme)
   }
 
   connectedCallback() {
     if (this.isDarkMode) {
-      document.documentElement.classList.add('dark')
-    }else {
-      document.documentElement.classList.remove('dark')
+      document.documentElement.classList.add("dark")
+    } else {
+      document.documentElement.classList.remove("dark")
     }
-    this.button?.addEventListener('click', this.toggleTheme.bind(this))
-    this.button?.addEventListener('keydown', this.toggleThemeWithKeyboard.bind(this))
-
+    this.button?.addEventListener("click", this.toggleTheme.bind(this))
+    this.button?.addEventListener("keydown", this.toggleThemeWithKeyboard.bind(this))
   }
 
   disconnectedCallback() {
-    this.button?.removeEventListener('click', this.toggleTheme)
+    this.button?.removeEventListener("click", this.toggleTheme)
     this.button?.removeEventListener("keydown", this.toggleThemeWithKeyboard)
   }
-
-
 
   /**
    * Check if mode is equal to 'dark'
@@ -52,11 +60,13 @@ export class HeadlessDarkMode extends HTMLElement {
   private toggleTheme(event: any): void {
     event.preventDefault()
     if (!this.isDarkMode) {
-      document.documentElement.classList.add('dark')
-      localStorage.theme = 'dark'
-    }else {
-      document.documentElement.classList.remove('dark')
-      localStorage.theme = 'light'
+      document.documentElement.classList.add("dark")
+      localStorage.theme = "dark"
+      this.dataset.theme = "dark"
+    } else {
+      document.documentElement.classList.remove("dark")
+      localStorage.theme = "light"
+      this.dataset.theme = "light"
     }
   }
 
@@ -65,8 +75,8 @@ export class HeadlessDarkMode extends HTMLElement {
    * @return void
    * */
   private toggleThemeWithKeyboard(event: any): void {
-    event.preventDefault()
     if (keycodeEquals(["Space", "Enter"], event)) {
+      event.preventDefault()
       this.toggleTheme(event)
     }
   }
@@ -78,6 +88,34 @@ export class HeadlessDarkMode extends HTMLElement {
   private get themeData(): string {
     return localStorage.theme
   }
+
+  /**
+   * Observe attributes that change
+   */
+  static get observedAttributes() {
+    return ["data-theme"]
+  }
+
+  /**
+   * Callback then attribute changed
+   * @param property
+   * @param oldValue
+   * @param newValue
+   */
+  attributeChangedCallback(property: any, oldValue: any, newValue: any) {
+    if (property === "data-theme" && oldValue !== newValue) {
+      this.updateAllDateThemeAttributes(newValue)
+    }
+  }
+
+  /**
+   * Update all attributes data-theme
+   * @param value
+   * @private
+   */
+  updateAllDateThemeAttributes(value: string) {
+    this.dataTheme.forEach((data) => (data.dataset.theme = value))
+  }
 }
 
-customElements.define('headless-darkmode', HeadlessDarkMode)
+customElements.define("headless-darkmode", HeadlessDarkMode)