+++ src/index.ts
@@ -2,7 +2,7 @@ import "./headless-popover"
 import "./headless-dropdown"
 import "./headless-navigation"
 import "./headless-disclosure"
-import "./headless-rating"
+// import "./headless-rating"
 import "./headless-darkmode"
 import "./headless-toggle"